+++ cmd/server.go
@@ -5,7 +5,6 @@ import (
 	"log"
 	"os"
 	"os/signal"
-	"strings"
 	"syscall"
 	"time"
 
@@ -57,7 +56,7 @@ func runServer(cmd *cobra.Command, args []string) {
 	log.Println("==== Starting gokakashi ====")
 	go startAPIServer(cfg)
 	go startWebServer(cfg)
-	logConfig(cfg)
+	log.Println(cfg)
 
 	<-done
 	log.Println("====  Exiting gokakashi. Bye! ==== ")
@@ -78,7 +77,6 @@ func startAPIServer(cfg *configv1.Config) {
 	configDB := restapiv1.InitDB(dbConfig)
 	defer configDB.Close()
 	db.RunMigrations(configDB)
-	// Populate the database
 	db.PopulateDatabase(configDB, cfg)
 
 	// ToDo: To be go routine who independently and routinely checks and assigns scans in agentTasks table
@@ -95,23 +93,4 @@ func startWebServer(cfg *configv1.Config) {
 	if err := webServer.ListenAndServe(); err != nil {
 		log.Fatalln("Error starting web server", err)
 	}
-}
-
-func logConfig(cfg *configv1.Config) {
-	log.Println()
-	log.Println("- - - - Configuration - - - - -")
-	log.Printf("  API Server URL: %s\n", cfg.APIServerURL())
-	if cfg.Site.LogAPITokenOnStartup {
-		log.Printf("  API Token: %s\n", cfg.Site.APIToken)
-	}
-	log.Println("")
-	log.Printf("  Web Server URL: %s\n", cfg.WebServerURL())
-	log.Println("")
-	log.Printf("  Database Host: %s\n", cfg.Database.Host)
-	log.Printf("  Database Port: %d\n", cfg.Database.Port)
-	log.Printf("  Database User: %s\n", cfg.Database.User)
-	log.Printf("  Database Name: %s\n", cfg.Database.Name)
-	log.Printf("  Database Password: %s\n", strings.Repeat("*", len(cfg.Database.Password)))
-	log.Println("- - - - - - - - - - - - - - - -")
-	log.Println()
 }
