+++ internal/config/v1/config.go
@@ -1,16 +1,10 @@
 package v1
 
 import (
-	"crypto/rand"
-	"encoding/base64"
 	"fmt"
-	"log"
-	"os"
-	"strconv"
+	"strings"
 
 	"github.com/shinobistack/gokakashi/ent/schema"
-
-	"gopkg.in/yaml.v2"
 )
 
 // Integration defines the configuration for external services
@@ -80,144 +74,6 @@ type Scanner struct {
 	Tool string `yaml:"tool"` // Example: Trivy, Synk, etc.
 }
 
-func LoadConfig(configFile string) (*Config, error) {
-	config := &Config{}
-
-	yamlFile, err := os.ReadFile(configFile)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read config file: %v", err)
-	}
-
-	err = yaml.Unmarshal(yamlFile, config)
-	if err != nil {
-		return nil, fmt.Errorf("error parsing YAML file: %v", err)
-	}
-
-	return config, nil
-}
-
-// validateConfig validates the loaded configuration to ensure required fields are present
-func ValidateConfig(config *Config) error {
-	if config.Site.APIToken == "" {
-		return fmt.Errorf("API token is missing")
-	}
-
-	// ToDo: Minimum one integration must be provided and that needs to be ...?
-	if len(config.Integrations) == 0 {
-		return fmt.Errorf("At least one integration must be defined")
-	}
-	for _, integration := range config.Integrations {
-		if integration.Name == "" || integration.Type == "" {
-			return fmt.Errorf("Integration name and type are required")
-		}
-	}
-	for _, policy := range config.Policies {
-		if policy.Name == "" {
-			return fmt.Errorf("Policy name is required")
-		}
-		if policy.Image.Registry == "" || policy.Image.Name == "" {
-			return fmt.Errorf("Policy image registry and name are required")
-		}
-		if policy.Trigger.Type == "cron" && policy.Trigger.Schedule == "" {
-			return fmt.Errorf("Policy with cron trigger must define a schedule")
-		}
-		for _, notify := range policy.Notify {
-			if notify.To == "" {
-				return fmt.Errorf("Notify 'to' field is required")
-			}
-			if notify.When == "" {
-				return fmt.Errorf("Notify 'when' field is required")
-			}
-		}
-	}
-	return nil
-
-	// ToDo: more validations to be added as per config design
-	// ToDo: How do we validated the token string for each webserver, currently this is not being used
-}
-
-func LoadAndValidateConfig(configPath string) (*Config, error) {
-	// Load the YAML configuration
-	log.Printf("Loading configuration from YAML file: %s", configPath)
-	cfg, err := LoadConfig(configPath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load config: %w", err)
-	}
-
-	// Validate the configuration
-	if err := ValidateConfig(cfg); err != nil {
-		return nil, fmt.Errorf("configuration validation failed: %w", err)
-	}
-	log.Println("Configuration loaded and validated successfully.")
-	return cfg, nil
-}
-
-func DefaultConfig() (*Config, error) {
-	apiToken := os.Getenv("GOKAKASHI_API_SERVER_TOKEN")
-	logAPITokenOnStartup := false
-	if apiToken == "" {
-		generatedToken, err := generateToken(32)
-		if err != nil {
-			return nil, fmt.Errorf("error generating an api token: %w", err)
-		}
-		apiToken = generatedToken
-		logAPITokenOnStartup = true
-	}
-
-	apiHost := os.Getenv("GOKAKASHI_API_SERVER_HOST")
-	apiPort, _ := strconv.Atoi(os.Getenv("GOKAKASHI_API_SERVER_PORT"))
-	if apiPort == 0 {
-		apiPort = 5556
-	}
-
-	webHost := os.Getenv("GOKAKASHI_WEB_SERVER_HOST")
-	webPort, _ := strconv.Atoi(os.Getenv("GOKAKASHI_WEB_SERVER_PORT"))
-	if webPort == 0 {
-		webPort = 5555
-	}
-
-	dbHost := os.Getenv("DB_HOST")
-	if dbHost == "" {
-		dbHost = "localhost"
-	}
-	dbPort, _ := strconv.Atoi(os.Getenv("DB_PORT"))
-	if dbPort == 0 {
-		dbPort = 5432
-	}
-	dbName := os.Getenv("DB_NAME")
-	if dbName == "" {
-		dbName = "postgres"
-	}
-	dbUser := os.Getenv("DB_USER")
-	if dbUser == "" {
-		dbUser = "postgres"
-	}
-	dbPassword := os.Getenv("DB_PASSWORD")
-	if dbPassword == "" {
-		dbPassword = "postgres"
-	}
-
-	return &Config{
-		Site: SiteConfig{
-			APIToken:             apiToken,
-			LogAPITokenOnStartup: logAPITokenOnStartup,
-			Host:                 apiHost,
-			Port:                 apiPort,
-		},
-		WebServer: WebServerConfig{
-			Host: webHost,
-			Port: webPort,
-		},
-		Database: DbConnection{
-			Host:     dbHost,
-			Port:     dbPort,
-			User:     dbUser,
-			Password: dbPassword,
-			Name:     dbName,
-		},
-	}, nil
-}
-
 func (c *Config) WebServerURL() string {
 	host := c.WebServer.Host
 	if host == "" {
@@ -234,13 +90,27 @@ func (c *Config) APIServerURL() string {
 	return fmt.Sprintf("http://%s:%d", host, c.Site.Port)
 }
 
-func generateToken(length int) (string, error) {
-	token := make([]byte, length)
+func (cfg *Config) String() string {
+	var s strings.Builder
 
-	_, err := rand.Read(token)
-	if err != nil {
-		return "", err
+	s.WriteString("\n")
+	s.WriteString("- - - - Configuration - - - - -")
+	s.WriteString("\n")
+
+	s.WriteString(fmt.Sprintf("  API Server URL: %s\n", cfg.APIServerURL()))
+	if cfg.Site.LogAPITokenOnStartup {
+		s.WriteString(fmt.Sprintf("  API Token: %s\n", cfg.Site.APIToken))
 	}
+	s.WriteString("\n")
+	s.WriteString(fmt.Sprintf("  Web Server URL: %s\n", cfg.WebServerURL()))
+	s.WriteString("\n")
+	s.WriteString(fmt.Sprintf("  Database Host: %s\n", cfg.Database.Host))
+	s.WriteString(fmt.Sprintf("  Database Port: %d\n", cfg.Database.Port))
+	s.WriteString(fmt.Sprintf("  Database User: %s\n", cfg.Database.User))
+	s.WriteString(fmt.Sprintf("  Database Name: %s\n", cfg.Database.Name))
+	s.WriteString(fmt.Sprintf("  Database Password: %s\n", strings.Repeat("*", len(cfg.Database.Password))))
+	s.WriteString("- - - - - - - - - - - - - - - -")
+	s.WriteString("\n")
 
-	return base64.URLEncoding.EncodeToString(token), nil
+	return s.String()
 }
